refactor /users/[id] page a bit more
